pyproject.toml
[project]
name = "missedcall-agent"
version = "0.1.0"
description = "Jan AI voicemail: ring Jan first with Twilio, then handle missed calls with Pipecat and send Slack recaps."
requires-python = ">=3.10"
dependencies = [
  "aiofiles>=24.1.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "loguru>=0.7.2",
  "pipecat-ai[deepgram,elevenlabs,openai,websocket]>=0.0.95",
  "pipecatcloud>=0.4.4",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.12",
  "twilio>=9.4.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
